% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{copy.conditional.column}
\alias{copy.conditional.column}
\title{Copies content from columns of one data.frame to another data.frame based column matching (conditional copy)}
\usage{
copy.conditional.column(x, y, columns.copy = names(y),
  columns.copy.x = columns.copy, columns.copy.y = columns.copy,
  by = intersect(names(x), names(y))[1], by.x = by, by.y = by)
}
\arguments{
\item{x}{destination data frame}

\item{y}{source data frame}

\item{columns.copy}{the columns from y that should be copied to x.}

\item{columns.copy.x}{the names of the new columns in x (default: columns.copy)}

\item{columns.copy.y}{the names of the columns in y (default: columns.copy)}

\item{by}{that matching column of x and y}

\item{by.x}{the name of the matching column in x (default: by)}

\item{by.y}{the name of the matching column in y (default: by)}
}
\value{
the data frame x with content from y in the columns (columns.copy) where by.x = by.y
}
\description{
Creates or replaces column content of the destination data frame (x) with data from (y) where the rows have matching content in column (by)
}
\examples{
test = data.frame(carb=c(1,2,3,4,5),desc=c('a','b','c','d','e'))
copy.conditional.column(mtcars, test, columns.copy = "desc", by = "carb")
}
