% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.init}
\alias{db.init}
\title{Initializes the database}
\usage{
db.init(host, user, password, dbname)
}
\arguments{
\item{host}{the IP address of the database}

\item{user}{the username}

\item{password}{the user's password}

\item{dbname}{the database's name you want to connect to}
}
\value{
void
}
\description{
Stores login information for later database connections.
}
\examples{
\dontrun{
db.init(host = '127.0.0.1',
        user = 'myUsername',
        password = 'myPassword',
        dbname='myDatabaseName')
}
}
