% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db.query}
\alias{db.query}
\title{Executes a SELECT-query}
\usage{
db.query(queryString)
}
\arguments{
\item{queryString}{the SELECT-query string}
}
\value{
the result set as data.frame
}
\description{
Executes a SELECT-query and returns the result set as a data.frame.
}
\section{Important}{
 db.init has to be called first!
}

\examples{
\dontrun{
participants = db.query("SELECT * FROM participants")
}
}
\seealso{
\code{\link{db.init}}
}
