% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstr.R
\name{effect.size.string}
\alias{effect.size.string}
\title{Provides a string representation of a effect size}
\usage{
effect.size.string(effect.size)
}
\arguments{
\item{effect.size}{the calculated effect size, e.g. r}
}
\value{
string representation of the effect size
}
\description{
< 0.1 = no effect, < 0.3 = small effect, < 0.5 = medium effect, otherwise = large effect
}
\examples{
effect.size.string(0.35)
}
