% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{env}
\alias{env}
\title{Setting up an environment to store the database login information}
\format{An object of class \code{environment} of length 4.}
\usage{
env
}
\description{
Setting up an environment to store the database login information
}
\keyword{internal}
