% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-string.R
\name{kruskal.wallis.string}
\alias{kruskal.wallis.string}
\title{Creates the statistic string for a Kruskal-Wallis test result}
\usage{
kruskal.wallis.string(kruskal.result)
}
\arguments{
\item{kruskal.result}{the kruskal-wallis test result}
}
\value{
string representation of the result
}
\description{
Format of the string is: H(df) = X^2, p < 0.05
}
\examples{
kruskal.wallis.string(kruskal.test(wt ~ carb,data=mtcars))
}
