% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-string.R
\name{mann.whitney.string}
\alias{mann.whitney.string}
\title{Creates the statistic string for a Wilcoxon-Mann-Whitney test result}
\usage{
mann.whitney.string(mann.result, sample.size = NULL)
}
\arguments{
\item{mann.result}{the Wilcoxon-Mann-Whitney test result}

\item{sample.size}{the sample's size. If provided, the effect-size (r) will be provided rather than z}
}
\value{
string representation of the result
}
\description{
Format of the string is: U = x, p < 0.05, z = -10
}
\examples{
mann.whitney.string(wilcox.test(wt ~ am, data=head(mtcars)))
}
