% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.factors.columns}
\alias{ml.pre.factors.columns}
\title{#########################################################
############## PRE PROCESSING 2 - FACTORS ###############
#########################################################
Identifies Factor columns}
\usage{
ml.pre.factors.columns(dataset)
}
\arguments{
\item{dataset}{dataset to investigate}
}
\value{
vector of column names that contain strings
}
\description{
Retrieves columns of a dataset that are strings and should be converted to factors.
}
\examples{
ml.pre.factors.columns(ml.testset())
}
