% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.factors.correct.data.frame}
\alias{ml.pre.factors.correct.data.frame}
\title{Converts strings to factors(data.frame)}
\usage{
ml.pre.factors.correct.data.frame(dataset)
}
\arguments{
\item{dataset}{dataset to apply the conversion on}
}
\value{
dataset with corrected factor values
}
\description{
Converts all string values in a data.frame to factor values.
}
\examples{
ml.pre.factors.correct.data.frame(ml.testset())
}
