% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.factors.correct}
\alias{ml.pre.factors.correct}
\title{Converts strings to factors}
\usage{
ml.pre.factors.correct(vec)
}
\arguments{
\item{vec}{column for string conversion (or data.frame)}
}
\value{
same column vector but with converted string values
}
\description{
Converts all string values of a column to factor values.
}
\examples{
ml.pre.factors.correct(ml.testset()$dep)
}
