% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.na.columns}
\alias{ml.pre.na.columns}
\title{#########################################################
############# PRE PROCESSING 1 - NA VALUES ##############
#########################################################
Identifies NA columns}
\usage{
ml.pre.na.columns(dataset)
}
\arguments{
\item{dataset}{dataset to investigate}
}
\value{
vector of column names that contain NA entries
}
\description{
Retrieves columns of a dataset that contains NA entries.
}
\examples{
ml.pre.na.columns(ml.testset())
}
