% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.na.correct.data.frame}
\alias{ml.pre.na.correct.data.frame}
\title{Replaces NA values(data.frame)}
\usage{
ml.pre.na.correct.data.frame(dataset, fun)
}
\arguments{
\item{dataset}{dataset to apply the replacement on}

\item{fun}{function to determine the replacement value. Is applied only on entries != NA.}
}
\value{
dataset with corrected NA values
}
\description{
Replaces all NA values in a data.frame with the value of a predefined method, such as the mean.
}
\examples{
ml.pre.na.correct.data.frame(ml.testset(),mean)
}
