% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.na.correct}
\alias{ml.pre.na.correct}
\title{Replaces NA values}
\usage{
ml.pre.na.correct(vec, fun = mean)
}
\arguments{
\item{vec}{column for NA replacement (or data.frame)}

\item{fun}{function to determine the replacement value. Is applied only on entries != NA.}
}
\value{
same column vector but with replaced NA values
}
\description{
Replaces NA values of a column with the value of a predefined method, such as mean.
}
\examples{
ml.pre.na.correct(ml.testset()$disp,mean)
}
