% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.scaling}
\alias{ml.pre.scaling}
\title{#########################################################
########## PRE PROCESSING 4 - FEATURE SCALING ###########
#########################################################
Scales columns' values}
\usage{
ml.pre.scaling(training_set, test_set = NULL, columns = NULL)
}
\arguments{
\item{training_set}{the training set or list(training_set=...,test_set=...)}

\item{test_set}{the test set or NULL}

\item{columns}{the columns that should be scaled as vector or NULL to scale all numeric columns}
}
\value{
list with 2 datasets: training_set and test_set
}
\description{
Applies Feature Scaling to columns
}
\examples{
ml.pre.scaling(ml.pre.split(ml.testset(),"dep"))
}
