% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.pre.split}
\alias{ml.pre.split}
\title{#########################################################
########### PRE PROCESSING 3 - SPLIT DATASET ############
#########################################################
Splits a dataset}
\usage{
ml.pre.split(dataset, depVariable, splitRatio = 0.8, seed = 123)
}
\arguments{
\item{dataset}{dataset to split}

\item{depVariable}{the dependent variable of the dataset (on which basis the split is done)}

\item{splitRatio}{ratio of the training_set size ]0,1]}

\item{seed}{the seed for the random number generator (or NULL for no change of the seed)}
}
\value{
list with 2 datasets: training_set and test_set
}
\description{
Splits a dataset into training_set and test_set.
}
\examples{
ml.pre.split(ml.testset(),"dep")
}
