% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.prepare}
\alias{ml.prepare}
\title{Prepares a dataset}
\usage{
ml.prepare(dataset, depColumn, missingDataFun = mean,
  transformFactors = TRUE, splitRatio = 0.8, seed = NULL,
  featureScaling = FALSE)
}
\arguments{
\item{dataset}{the dataset to prepare}

\item{depColumn}{the column name that contains the dependent variable}

\item{missingDataFun}{function to replace NA entries or NULL to skip NA correction}

\item{transformFactors}{set to true to convert character columns into factors}

\item{splitRatio}{the amount of data that should go into the training set (80\% = 0.8)}

\item{seed}{if results should be done consistently random, enter a seed number here, otherwise NULL}

\item{featureScaling}{set to true to automatically apply feature scaling to every numeric column}
}
\value{
list with 2 datasets: training_set and test_set
}
\description{
Prepares a dataset for machine learning operations
}
\examples{
ml.prepare(ml.testset())
}
