% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml.testset}
\alias{ml.testset}
\title{#########################################################
######################### TESTS #########################
#########################################################
Testset for ML functions}
\usage{
ml.testset()
}
\value{
data.frame object for testing
}
\description{
Creates a testset based on mtcars for this methods to test.
Contains a column 'dep' that describes the following relationship:
if cyl == 4 OR qsec < 16,
then ml.testset()$dep = 'class1',
otherwise ml.testset()$dep = 'class2'
}
\examples{
ml.testset()
}
