% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstr.R
\name{p.value.level}
\alias{p.value.level}
\title{Classifies the p.value into one of 4 significance classes}
\usage{
p.value.level(p.value)
}
\arguments{
\item{p.value}{p-value}
}
\value{
string representation of the p-value class
}
\description{
ns, p < 0.05, p < 0.01, p < 0.001
}
\examples{
p.value.level(0.03)
}
