% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{p.value.string}
\alias{p.value.string}
\title{Provides a string representation of a p-value}
\usage{
p.value.string(p.value)
}
\arguments{
\item{p.value}{p-value}
}
\value{
string representation of the p-value
}
\description{
< 0.1 = ., < 0.05 = *, < 0.01 = **, < 0.001 = ***
}
\examples{
p.value.string(0.95)
}
