% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{reset.output}
\alias{reset.output}
\title{Closes all open sinks and graphic devices}
\usage{
reset.output()
}
\value{
void
}
\description{
Closes the sinks that are still open and closes all graphic devices
}
\examples{
reset.output()
}
