% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save.plot}
\alias{save.plot}
\title{Stores a ggplot}
\usage{
save.plot(p, filename, ppi = 300, paper = "a4r", width = 11.69,
  height = 8.27, print.default.graphicdevice = TRUE)
}
\arguments{
\item{p}{the ggplot-plot}

\item{filename}{filename of the file WITH extension, which describes the format. Currently supported is "pdf", "png", "jpeg", "bmp"}

\item{ppi}{the resultion of the picture (not used important for pdf)}

\item{paper}{default paper - only important for pdf (see pdf(paper=?))}

\item{width}{width of the ploting area in inch (width * ppi defines the size of the picture formats)}

\item{height}{height of the ploting area in inch (as for width)}

\item{print.default.graphicdevice}{when true, the plot will be plotted on the default graphics device as well}
}
\value{
void
}
\description{
Creates a DinA4 document (on default) and places the plot on the whole document
}
\examples{
save.plot(statistics.plot.frequency(mtcars, c("cyl","carb")),"test.pdf")
}
