% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics.descriptive.nonparametric}
\alias{statistics.descriptive.nonparametric}
\title{Calculates the descriptive statistic of non-parametric data}
\usage{
statistics.descriptive.nonparametric(data, column.dependent,
  columns.independent)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable}

\item{columns.independent}{a vector of columns that contain the independent variables}
}
\description{
Returns a data.frame with the following values
\itemize{
\item{N - the sample size}
\item{Sum - the sample size}
\item{Percentiles.25 - 25th percentiles}
\item{Percentiles.50 - 50th (median) percentiles}
\item{Percentiles.50 - 75th percentiles}
\item{Min - the sample's minimum}
\item{Max - the sample's maximum}
}
}
\examples{
statistics.descriptive.nonparametric(mtcars, "carb", c("cyl"))
}
