% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-descriptive.R
\name{statistics.descriptive.parametric}
\alias{statistics.descriptive.parametric}
\title{Calculates the descriptive statistic of parametric data}
\usage{
statistics.descriptive.parametric(data, column.dependent, columns.independent,
  conf.interval = 0.95, conf.accurate = TRUE)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{conf.interval}{the confidence interval desired (default 95\%)}

\item{conf.accurate}{TRUE to use the accurate confidence interval. This means, that the CI.factor is calculated using the sample size - 1. Otherwise 10.000 is assumed as sample size.}
}
\value{
a data frame containing basic statistic values
}
\description{
Returns a data.frame with the following values
\itemize{
\item{N - the sample size}
\item{Sum - sum of all sample's values}
\item{Mean - the sample's mean}
\item{SS - the squared sum of errors}
\item{Var - the sample's variance}
\item{SD - the sample's standard deviation}
\item{SE - the estimated population standard error}
\item{CI.Lower - lower confidence interval}
\item{CI.Upper - upper confidence interval}
\item{CI.Factor - confidence interval factor}
\item{CI.Range - confidence interval range (CI.Factor * SE)}
\item{Min - the sample's minimum}
\item{Max - the sample's maximum}
}
}
\examples{
statistics.descriptive.parametric(mtcars, "wt", c("cyl"),
                                  conf.interval = .95,
                                  conf.accurate = TRUE)
}
