% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-descriptive.R
\name{statistics.descriptive}
\alias{statistics.descriptive}
\title{Calculates the descriptive statistic of parametric data}
\usage{
statistics.descriptive(data, column.dependent, columns.independent,
  data.type = "parametric", conf.interval = 0.95, conf.accurate = TRUE)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable | will be ignored when data.type == 'frequency'}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{data.type}{the data-type. Allowed values are 'parametric', 'nonparametric', and 'frequency'}

\item{conf.interval}{the confidence interval desired (default 95\%) | only used when data.type == 'parametric'}

\item{conf.accurate}{TRUE to use the accurate confidence interval. This means, that the CI.factor is calculated using the sample size - 1. Otherwise 10.000 is assumed as sample size.  | only used when data.type == 'parametric'}
}
\value{
a data frame containing basic statistic values
}
\description{
Returns a data.frame. The data.frame data depends on the type of data.
}
\examples{
statistics.descriptive(mtcars, "carb", c("cyl"), data.type = 'nonparametric')
}
\seealso{
\code{\link{statistics.descriptive.parametric}}

\code{\link{statistics.descriptive.nonparametric}}

\code{\link{statistics.descriptive.frequency}}
}
