% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-plot.R
\name{statistics.histogram}
\alias{statistics.histogram}
\title{Creates histograms of the data}
\usage{
statistics.histogram(data, column.dependent, columns.independent,
  column.fill = NULL, bin.width = NULL, bins = 30, title = TRUE)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable (x-axis)}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{column.fill}{column for a stacked histogram}

\item{bin.width}{width of each bin}

\item{bins}{how many bins you want to have (will be ignored if bin.width is set)}

\item{title}{true to get a title printed on each plot with the current condition}
}
\value{
a data frame containing basic statistic values
}
\description{
A list of histogram ggplot-plots
}
\examples{
statistics.histogram(mtcars, "carb", c("cyl"), column.fill = "gear", bins = 5)
}
\seealso{
\code{\link{statistics.descriptive.parametric}}

\code{\link{statistics.descriptive.nonparametric}}

\code{\link{statistics.descriptive.frequency}}
}
