% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics.plot.nonparametric}
\alias{statistics.plot.nonparametric}
\title{Creates descriptive plot of non parametric data}
\usage{
statistics.plot.nonparametric(data, column.dependent, columns.independent,
  xlab = NULL, ylab = NULL, legend.lab = NULL, xlim = NULL,
  ylim = NULL, title = NULL)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable (x-axis)}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{xlab}{label for x-axis (columns.independent[1] if provided)}

\item{ylab}{label for y-axis (column.dependent)}

\item{legend.lab}{label for legend (columns.independent[2] if provided)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{title}{title for the graph}
}
\value{
a ggplot-plot
}
\description{
Creates a boxplot as ggplot-plot
}
\examples{
statistics.plot.nonparametric(mtcars, "mpg",c("cyl","carb"))
}
