% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-plot.R
\name{statistics.plot.parametric}
\alias{statistics.plot.parametric}
\title{Creates descriptive plot of parametric data}
\usage{
statistics.plot.parametric(data, column.dependent, columns.independent,
  xlab = NULL, ylab = NULL, legend.lab = NULL, xlim = NULL,
  ylim = NULL, title = NULL, position = position_dodge(),
  conf.interval = 0.95, conf.accurate = TRUE)
}
\arguments{
\item{data}{the data as a data.frame}

\item{column.dependent}{the name of the column that contains the dependent variable (x-axis)}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{xlab}{label for x-axis (columns.independent[1] if provided)}

\item{ylab}{label for y-axis (column.dependent)}

\item{legend.lab}{label for legend (columns.independent[2] if provided)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{title}{title for the graph}

\item{position}{the arrangement of bars in the bar plot (dodge on default, use "stack" for stacked version)}

\item{conf.interval}{the confidence interval desired (default 95\%)}

\item{conf.accurate}{TRUE to use the accurate confidence interval. This means, that the CI.factor is calculated using the sample size - 1. Otherwise 10.000 is assumed as sample size.}
}
\value{
a ggplot-plot
}
\description{
Creates first a parametric descriptive statistic and plots it into a ggplot-plot
}
\examples{
statistics.plot.parametric(mtcars, "wt",c("cyl","carb"))
}
