% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstr.R
\name{wilcox.effect.size}
\alias{wilcox.effect.size}
\title{Provides the effect size for a wilcoxon test}
\usage{
wilcox.effect.size(wilcox.result, sample.size)
}
\arguments{
\item{wilcox.result}{the wilcoxon test result}

\item{sample.size}{the sample's size}
}
\value{
r-value of the test result
}
\description{
r = wilcoxon.z / sqrt(sample.size)
}
\examples{
wilcox.effect.size(wilcox.test(wt ~ am, data=head(mtcars)),nrow(head(mtcars)))
}
