% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstr.R
\name{wilcox.z}
\alias{wilcox.z}
\title{Converts the wilcox test's p.value into a z-value}
\usage{
wilcox.z(wilcox.result)
}
\arguments{
\item{wilcox.result}{the wilcoxon test result}
}
\value{
z-value of the test result
}
\description{
qnorm(wilcox.result$p.value / 2), based on Andy Field's formular
}
\examples{
wilcox.z(wilcox.test(wt ~ am, data=head(mtcars)))
}
