% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save.plot}
\alias{save.plot}
\title{Stores a ggplot}
\usage{
save.plot(p, filename, dpi = 300, paper = "special", width = 11.69,
  height = 8.27, print.default.graphicdevice = TRUE)
}
\arguments{
\item{p}{the ggplot-plot or a list of ggplots. Filename will be adjusted with increasing numbers, except for pdf files. These will contain one page per plot.}

\item{filename}{filename of the file WITH extension, which describes the format. Currently supported are "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg"}

\item{dpi}{the resultion of the picture (important for raster graphics)}

\item{paper}{default paper (currently specified by width and height) - only important for pdfs (see pdf(paper=?))}

\item{width}{width of the ploting area in inch (width * ppi defines the size of the picture formats)}

\item{height}{height of the ploting area in inch (as for width)}

\item{print.default.graphicdevice}{when true, the plot will be plotted on the default graphics device as well}
}
\value{
void
}
\description{
Creates a DinA4 document (on default) and places the plot on the whole document.
}
\examples{
save.plot(statistics.plot.frequency(mtcars, c("cyl","carb")),"test.pdf")
}
