% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-plot.R
\name{statistics.plot.frequency}
\alias{statistics.plot.frequency}
\title{Creates descriptive plot of frequency data}
\usage{
statistics.plot.frequency(data, columns.independent, drop = FALSE,
  xlab = NULL, ylab = NULL, legend.lab = NULL, xlim = NULL,
  ylim = NULL, title = NULL, position = position_dodge())
}
\arguments{
\item{data}{the data as a data.frame}

\item{columns.independent}{a vector of columns that contain the independent variables}

\item{drop}{should combinations of variables that do not appear in the columns.independent be preserved or dropped (default = FALSE)}

\item{xlab}{label for x-axis (columns.independent[1] if provided)}

\item{ylab}{label for y-axis (frequency)}

\item{legend.lab}{label for legend (columns.independent[2] if provided)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{title}{title for the graph}

\item{position}{the arrangement of bars in the bar plot (dodge on default, use "stack" for stacked version)}
}
\value{
a ggplot-plot
}
\description{
Creates a barchart as ggplot-plot
}
\examples{
statistics.plot.frequency(mtcars, c("cyl","carb"))
}
