% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-string.R
\name{wilcox.effect.size}
\alias{wilcox.effect.size}
\title{Provides the effect size for a wilcoxon test}
\usage{
wilcox.effect.size(sample.size, wilcox.result = NULL, z.value = NULL)
}
\arguments{
\item{sample.size}{the sample's size}

\item{wilcox.result}{the wilcoxon test result (only used if z.value == NULL otherwise z.value is used)}

\item{z.value}{the wilcoxon test's z-value}
}
\value{
r-value of the test result
}
\description{
r = wilcoxon.z / sqrt(sample.size)
}
\examples{
wilcox.effect.size(nrow(head(mtcars)), z.value = -1.644854)
wilcox.effect.size(nrow(head(mtcars)), wilcox.result = wilcox.test(wt ~ am, data=head(mtcars)))
}
